llm-based-extraction/requirements.txt
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
pyyaml>=6.0
